src/requirements.txt
aiohttp==3.13.3
discord.py==2.6.4
pillow==12.1.0
pydantic==2.12.5
python-dotenv==1.2.1
tenacity==9.1.4
diskcache==5.6.3
aiosqlite==0.22.1
